Allowing 'as' prop in Link and fixing 'href' prop
Previously, if a user already passed in an 'as' prop to the i18n Link
component, it would overwrite the 'as' prop to the next/link Link component.
This checks to see if an 'as' prop has been passed in to the i18n link
component and, if so, modifies it, rather than overwriting it.

Also, if the 'href' prop happened to already have query parameters, the 'lng'
parameter would be incorrectly added (e.g. if 'href=/foo?bar', then it would
be modified to 'href=/foo?bar?lng=en').  This commit also checks to see if
query parameters already exist and if, so, appends the lng parameter properly.

Lastly, this commit adds some initial test infrastructure using enzyme and
jest with code coverage and adds tests for the Link component.